+++ lib/lightning/policies/project_users.ex
@@ -5,10 +5,29 @@ defmodule Lightning.Policies.ProjectUsers do
   @behaviour Bodyguard.Policy
 
   alias Lightning.Projects
-  alias Lightning.Projects.Project
+  alias Lightning.Projects.{ProjectUser, Project}
   alias Lightning.Accounts.User
 
-  @type actions :: :edit_jobs
+  @type actions ::
+          :create_workflow
+          | :edit_jobs
+          | :create_job
+          | :delete_job
+          | :run_job
+          | :rerun_job
+          | :view_last_job_inputs
+          | :view_workorder_input
+          | :view_workorder_run
+          | :view_workorder_history
+          | :view_project_name
+          | :view_project_description
+          | :view_project_credentials
+          | :view_project_collaborators
+          | :delete_project
+          | :edit_digest_alerts
+          | :edit_project_name
+          | :edit_project_description
+          | :add_project_collaborator
 
   @doc """
   authorize/3 takes an action, a user, and a project. It checks the user's role
@@ -25,7 +44,50 @@ defmodule Lightning.Policies.ProjectUsers do
           Lightning.Accounts.User.t(),
           Lightning.Projects.Project.t()
         ) :: boolean
-  def authorize(:edit_jobs, %User{} = user, %Project{} = project) do
-    Projects.get_project_user_role(user, project) in [:admin, :editor, :owner]
-  end
+  def authorize(:delete_project, %User{} = user, %Project{} = project),
+    do: Projects.get_project_user_role(user, project) in [:owner]
+
+  def authorize(
+        :edit_digest_alerts,
+        %User{id: id} = _user,
+        %ProjectUser{user_id: user_id} = _project
+      ),
+      do: id == user_id
+
+  def authorize(action, %User{} = user, %Project{} = project)
+      when action in [
+             :create_workflow,
+             :edit_jobs,
+             :create_job,
+             :delete_job,
+             :run_job,
+             :rerun_job
+           ],
+      do:
+        Projects.get_project_user_role(user, project) in [
+          :owner,
+          :admin,
+          :editor
+        ]
+
+  def authorize(action, %User{} = user, %Project{} = project)
+      when action in [
+             :edit_project_name,
+             :edit_project_description,
+             :add_project_collaborator
+           ],
+      do: Projects.get_project_user_role(user, project) in [:owner, :admin]
+
+  def authorize(action, %User{} = user, %Project{} = project)
+      when action in [
+             :view_last_job_inputs,
+             :view_workorder_input,
+             :view_workorder_run,
+             :view_workorder_history,
+             :view_project_name,
+             :view_project_description,
+             :view_project_credentials,
+             :view_project_collaborators
+           ],
+      do: Projects.is_member_of?(project, user)
 end